requirements.txt
Django==2.2.6
djangorestframework==3.10.3
request==2019.4.13
requests==2.22.0
urllib3==1.25.6
utils==0.9.0
detectlanguage==1.3.0
